Update fan of ADC detect voltage by uint16_t is correct uint8_t is error .

+++ Core/Inc/bsp_control.h
@@ -35,7 +35,7 @@ typedef struct {
    //adc 
 	uint8_t ADC_channel_No;
 	uint16_t ptc_temp_voltage;
-	uint8_t fan_detect_voltage;
+	uint16_t fan_detect_voltage;
 
 	uint8_t  gPlasma;
     uint8_t  gPtc;
